home css style add

+++ src/components/Crow.js
@@ -1,20 +1,37 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { dbService, storageService } from 'firebaseSetup';
 import { COLLECTION } from '../constants';
+import { Form, Button } from 'react-bootstrap';
+import { faPen, faTrashAlt } from "@fortawesome/free-solid-svg-icons";
+import LoadingButton from '../entities/Button/LoadingButton';
+import './Crow.scss';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 const Crow = ({ crowObject, isOwner }) => {
   const userDoc = `${COLLECTION}/${crowObject.id}`;
   const [isEditing, setIsEditing] = useState(false);
   const [newCrow, setNewCrow] = useState(crowObject.text);
+  const [isSubmitting, setIsSubmitting] = useState(false);
+
+  useEffect(() => {
+    return () => {
+      setIsEditing(false);
+      setNewCrow(null);
+      setIsSubmitting(false);
+    }
+  }, []);
 
   const toggleEditing = () => setIsEditing(prev => !prev);
 
   const onDeleteClick = async () => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     const ok = window.confirm('삭제하시겠습니까악?');
     if (ok) {
       await dbService.doc(userDoc).delete();
       await storageService.refFromURL(crowObject.imgUrl).delete();
     }
+    setIsSubmitting(false);
   };
 
   const onEditChange = (event) => {
@@ -28,40 +45,66 @@ const Crow = ({ crowObject, isOwner }) => {
   }
 
   const onSubmit = async (event) => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     event.preventDefault();
     await dbService.doc(userDoc).update({
       text: newCrow,
     });
     setIsEditing(false);
+    setIsSubmitting(false);
   }
 
   return (
     <div>
       { isEditing ? (
         <>
-          <form onSubmit={onSubmit}>
-            <input
+          <Form
+            className="crow-card-fixed"
+            onSubmit={onSubmit}
+          >
+            <Form.Control
               type="text"
               value={newCrow}
               required
               placeholder="글을 작성해라 까악~"
               onChange={onEditChange}
             />
-            <button>수정</button>
-            <button type="button" onClick={onEditCancle}>취소</button>
-          </form>
+            <LoadingButton
+              type="submit"
+              isLoading={isSubmitting}
+              className="crow-card-fixed__modify"
+            >수정</LoadingButton>
+            <Button
+              type="button"
+              onClick={onEditCancle}
+              className="crow-card-fixed__cancle"
+            >취소</Button>
+          </Form>
         </>
       ) : (
-        <>
-          <h4>{crowObject.text}</h4>
+        <div className="crow-card">
+          { isOwner &&
+            <div className="crow-card--btn-group">
+              <Button
+                type="button"
+                className="crow-card__btn-edit"
+                variant=""
+                onClick={toggleEditing}
+              ><FontAwesomeIcon icon ={faPen} /></Button>
+              <LoadingButton
+                type="button"
+                className="crow-card__btn-delete"
+                buttonClick={onDeleteClick}
+                isLoading={isSubmitting}
+              ><FontAwesomeIcon icon ={faTrashAlt} /></LoadingButton>
+            </div>
+          }
+          <span className="crow-card--text">{crowObject.text}</span>
           { crowObject.imgUrl &&
             <img src={crowObject.imgUrl} width="50px" height="50px" />
           }
-          { isOwner && <>
-            <button onClick={onDeleteClick}>Delete Crow</button>
-            <button onClick={toggleEditing}>Edit Crow</button>
-          </> }
-        </>
+        </div>
       )}
     </div>
   );
